+++ src/app/GameSession.cpp
@@ -3,6 +3,7 @@
 #include "app/AudioMixPolicy.h"
 #include "app/ChartAudioPlayback.h"
 #include "app/ChartAudioStreaming.h"
+#include "app/ChartFileHash.h"
 #include "app/MemoryDiagnostics.h"
 #include "audio/OggVorbisDecoder.h"
 
@@ -44,6 +45,7 @@
 #include "app/RuntimeConfigMigration.h"
 #include "app/PersistedRuntimeConfig.h"
 #include "app/PeerBattleRules.h"
+#include "app/ReplayVerifier.h"
 #include "app/SessionResultStatus.h"
 #include "config/Keymap.h"
 #include "game/SpeedManager.h"